LocationSelectViewModelをインタフェースと実装に分ける
